DDD-23 Refactor package path
